pyproject.toml
[tool.poetry]
name = "aucpi"
version = "0.1.0"
description = "Adjusts Australian dollars for inflation."
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.7.1,<4.0"
appdirs = "^1.4.4"
pandas = "^1.3.2"
openpyxl = "^3.0.7"
xlrd = "^2.0.1"
cached-property = "^1.5.2"
python-dateutil = "^2.8.2"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.2.0"
pytest = "^6.2.4"
coverage = "^5.5"

[build-system]
requires = ["setuptools >= 40.6.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aucpi = 'aucpi.main:app'